requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0

